Break down scanning.js into multiple files
* New scanningHooks.js and scanningRequests.js files to hold the logic for hooks and requests related to scanning
* Rename scanning.js into scanningLib.js
