+++ src/scanning/scanningOptions/ScanningOptionsDiscarding/ScanningOptionsDiscarding.jsx
@@ -20,7 +20,7 @@ import { ControlledRadioGroup } from "../../../common/ControlledRadioGroup/Contr
  * @param {Partial<import("react-hook-form").FieldErrorsImpl<import("react-hook-form").DeepRequired<import("react-hook-form").FieldValues>>> & {root?: Record<string, import("react-hook-form").GlobalError> & import("react-hook-form").GlobalError}} props.errors
  * @param {import("react-hook-form").Control<any,any>} props.control
  * @param {import("react-hook-form").UseFormWatch<any>} props.watch
- * @param {import("../../scanning.js").Group[]} props.userAccessibleGroups
+ * @param {import("../../scanningLib.js").Group[]} props.userAccessibleGroups
  * @param {React.MutableRefObject<any>} props.modal
  * @returns {JSX.Element}
  */
@@ -32,7 +32,7 @@ export const ScanningOptionsDiscarding = ({
   userAccessibleGroups,
   modal,
 }) => {
-  /** @type {import("../../scanning.js").Group[]} */
+  /** @type {import("../../scanningLib.js").Group[]} */
   const junkGroups = watch("junkGroups").map(
     (/** @type {string[]} */ groupId) =>
       userAccessibleGroups.find((group) => group.id === +groupId),
@@ -49,7 +49,7 @@ export const ScanningOptionsDiscarding = ({
           .map((/** @type {string} */ groupId) =>
             userAccessibleGroups.find((group) => group.id === +groupId),
           )
-          .map((/** @type {import("../../scanning.js").Group} */ group) => (
+          .map((/** @type {import("../../scanningLib.js").Group} */ group) => (
             <IonChip key={group.id}>{group.name}</IonChip>
           ))}
         <IonModal
@@ -121,7 +121,7 @@ export const ScanningOptionsDiscarding = ({
                 >
                   {junkGroups.map(
                     (
-                      /** @type {import("../../scanning.js").Group} */ group,
+                      /** @type {import("../../scanningLib.js").Group} */ group,
                     ) => (
                       <IonSelectOption key={group.id} value={group.id}>
                         {group.name}
